Refactored all files and folders
