+++ src/asset_loader/mod.rs
@@ -1,0 +1,23 @@
+use crate::GameState;
+use bevy::prelude::*;
+use bevy_asset_loader::prelude::*;
+
+pub mod resources;
+
+use resources::*;
+
+pub struct AssetLoaderPlugin;
+
+/// This plugin loads all assets using [`AssetLoader`] from a third party bevy plugin
+/// Alternatively you can write the logic to load assets yourself
+/// If interested, take a look at <https://bevy-cheatbook.github.io/features/assets.html>
+impl Plugin for AssetLoaderPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_loading_state(
+            LoadingState::new(GameState::Loading).continue_to_state(GameState::Menu),
+        )
+        .add_collection_to_loading_state::<_, FontAssets>(GameState::Loading)
+        .add_collection_to_loading_state::<_, AudioAssets>(GameState::Loading)
+        .add_collection_to_loading_state::<_, TextureAssets>(GameState::Loading);
+    }
+}
