+++ src/audio/resources.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::*;
+
+#[derive(Resource)]
+pub struct BackgroundMusicAudio(Handle<AudioSource>);
