+++ src/camera/systems.rs
@@ -1,0 +1,11 @@
+use bevy::prelude::*;
+use bevy::window::PrimaryWindow;
+
+pub fn setup_camera(mut commands: Commands, window_query: Query<&Window, With<PrimaryWindow>>) {
+    let window = window_query.get_single().unwrap();
+
+    commands.spawn(Camera2dBundle {
+        transform: Transform::from_xyz(window.width() / 2.0, window.height() / 2.0, 1000.0),
+        ..default()
+    });
+}
