+++ src/enemy/components.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::*;
+
+#[derive(Component, Debug)]
+pub struct Enemy;
