+++ src/enemy/events.rs
@@ -1,0 +1,2 @@
+#[derive(Default)]
+pub struct EnemyHitWallEvent;
