+++ src/events.rs
@@ -1,0 +1,4 @@
+#[derive(Default)]
+pub struct GameOverEvent {
+    pub score: u32,
+}
