+++ src/physics/systems.rs
@@ -1,0 +1,18 @@
+use bevy::prelude::*;
+use bevy_rapier2d::prelude::*;
+
+#[derive(Component)]
+pub struct Walls;
+
+pub fn display_collision_events(
+    mut collision_events: EventReader<CollisionEvent>,
+    mut contact_force_events: EventReader<ContactForceEvent>,
+) {
+    for collision_event in collision_events.iter() {
+        println!("Received collision event: {:?}", collision_event);
+    }
+
+    for contact_force_event in contact_force_events.iter() {
+        println!("Received contact force event: {:?}", contact_force_event);
+    }
+}
