+++ src/player/events.rs
@@ -1,0 +1,5 @@
+#[derive(Default)]
+pub struct PlayerHitEnemyEvent;
+
+#[derive(Default)]
+pub struct PlayerStarPickupEvent;
