+++ src/resources.rs
@@ -1,0 +1,12 @@
+use bevy::prelude::*;
+
+#[derive(Resource)]
+pub struct Score {
+    pub value: u32,
+}
+
+impl Default for Score {
+    fn default() -> Self {
+        Self { value: 0 }
+    }
+}
