+++ src/star/resources.rs
@@ -1,0 +1,16 @@
+use bevy::prelude::*;
+
+const STAR_SPAWN_TIME: f32 = 5.0;
+
+#[derive(Resource)]
+pub struct StarSpawnTimer {
+    pub timer: Timer,
+}
+
+impl Default for StarSpawnTimer {
+    fn default() -> Self {
+        Self {
+            timer: Timer::from_seconds(STAR_SPAWN_TIME, TimerMode::Repeating),
+        }
+    }
+}
