+++ src/states.rs
@@ -1,0 +1,15 @@
+// This example game uses States to separate logic
+use bevy::prelude::*;
+
+// See https://bevy-cheatbook.github.io/programming/states.html
+// Or https://github.com/bevyengine/bevy/blob/main/examples/ecs/state.rs
+#[derive(States, Default, Clone, Eq, PartialEq, Debug, Hash)]
+pub enum GameState {
+    // During the loading State the AssetLoaderPlugin will load our assets
+    #[default]
+    Loading,
+    // During this State the actual game logic is executed
+    Playing,
+    // Here the menu is drawn and waiting for player interaction
+    Menu,
+}
