+++ src/walls/systems.rs
@@ -1,0 +1,36 @@
+use super::components::*;
+use bevy::{prelude::*, window::PrimaryWindow};
+use bevy_rapier2d::prelude::*;
+
+pub const WALLS_MARGIN: f32 = 5.0;
+
+// This system is responsible for setting up the play area pyhsics bounds
+pub fn spawn_walls(mut commands: Commands, window_query: Query<&Window, With<PrimaryWindow>>) {
+    let window = window_query.get_single().unwrap();
+
+    // Create the play area bounds
+    let play_area_width = window.width() - WALLS_MARGIN;
+    let play_area_height = window.height() - WALLS_MARGIN;
+
+    commands
+        .spawn(SpatialBundle::default())
+        .insert(RigidBody::Fixed)
+        .insert(Restitution {
+            coefficient: 1.0,
+            combine_rule: CoefficientCombineRule::Min,
+        })
+        .insert(Friction {
+            coefficient: 0.0,
+            combine_rule: CoefficientCombineRule::Min,
+        })
+        .insert(Collider::polyline(
+            vec![
+                Vect::new(WALLS_MARGIN, WALLS_MARGIN),
+                Vect::new(play_area_width, WALLS_MARGIN),
+                Vect::new(play_area_width, play_area_height),
+                Vect::new(WALLS_MARGIN, play_area_height),
+            ],
+            Some(vec![[0, 1], [1, 2], [2, 3], [3, 0]]),
+        ))
+        .insert(Walls);
+}
